layout/style/test/test_custom_content_inheritance.html

<!doctype html>
<title>Test for custom content inheritance</title>
<style>
  html { color: red !important; }
</style>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script>
onload = function() {
  try {
    let doc = SpecialPowers.wrap(document);
    let div = doc.createElement('div');
    div.id = "test-id";
    ok(!!doc.insertAnonymousContent,
       "Must have the insertAnonymousContent API");
    let content = doc.insertAnonymousContent(div);
    ok(!!content, "Must have anon content");
    isnot(content.getComputedStylePropertyValue("test-id", "color"),
          getComputedStyle(document.documentElement).color,
          "Custom anon content shouldn't inherit from the root element");
  } catch(e) {
    ok(false, "Threw: " + e);
  }
  SimpleTest.finish();
};
SimpleTest.waitForExplicitFinish();
</script>
